@@ -28,7 +28,7 @@
     "intl": "1.2.5",
     "jquery": "3.1.1",
     "reflect-metadata": "0.1.9",
-    "rxjs": "5.1.0",
+    "rxjs": "5.4.3",
     "ts-helpers": "1.1.2",
     "web-animations-js": "^2.3.1",
     "zone.js": "0.8.4"
@@ -49,7 +49,7 @@
     "protractor": "5.1.0",
     "ts-node": "2.1.0",
     "tslint": "4.5.0",
-    "typescript": "2.2.0",
+    "typescript": "2.5.2",
     "webdriver-manager": "10.2.5"
   }
 }
