@@ -1,1 +1,3 @@
-export const MEAT_API ='http://localhost:3000';
+import { environment } from "../environments/environment";
+
+export const MEAT_API = environment.api
