@@ -1,14 +1,28 @@
 import { Component, OnInit, Input, EventEmitter, Output } from '@angular/core';
+import { trigger, state, style, transition, animate } from '@angular/animations';
 import { MenuItem } from './menu-item.model';
 
 @Component({
   selector: 'mt-menu-item',
-  templateUrl: './menu-item.component.html'
+  templateUrl: './menu-item.component.html',
+  animations: [
+    trigger('menuItem-appeared', [
+      state('ready', style ({
+        opacity: 1
+      })),
+      transition('void => ready', [
+        style({opacity:0, transform: 'translateY(-20px)'}),
+        animate('300ms 0s ease-in')
+      ])
+    ])
+  ]
 })
 export class MenuItemComponent implements OnInit {
 
   @Input() menuItem: MenuItem;
   @Output() add = new EventEmitter();
+
+  menuItemState = 'ready';
 
   constructor() { }
 
