@@ -1,12 +1,26 @@
 import { Component, OnInit, Input } from '@angular/core';
+import { trigger, state, style, transition, animate } from '@angular/animations';
 
 import { Restaurant } from './restaurant.model';
 
 @Component({
   selector: 'mt-restaurant',
-  templateUrl: './restaurant.component.html'
+  templateUrl: './restaurant.component.html',
+  animations: [
+    trigger('restaurant-appeared', [
+      state('ready', style ({
+        opacity: 1
+      })),
+      transition('void => ready', [
+        style({opacity:0, transform: 'translate(-30px, -10px)'}),
+        animate('300ms 0s ease-in-out')
+      ])
+    ])
+  ]
 })
 export class RestaurantComponent implements OnInit {
+
+  restaurantState = 'ready';
 
   @Input() restaurant: Restaurant;
   
