@@ -6,7 +6,7 @@ export interface Restaurant {
     deliveryEstimate: string;
     rating: number;
     imagePath: string;
-    about: string;
-    hours: string;
+    about?: string;
+    hours?: string;
 
 }
