@@ -7,7 +7,7 @@ import { RestaurantsService } from './restaurants.service';
 @Component({
   selector: 'mt-restaurants',
   templateUrl: './restaurants.component.html'
-})
+ })
 
 export class RestaurantsComponent implements OnInit {
 
